fix: added core library, main functionality have separated
