+++ src/LearnApp.WebAPI/Startup.cs
@@ -1,9 +1,7 @@
-using LearnApp.WebAPI.Models;
+using LearnApp.Core.Models;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json.Serialization;
 
 namespace LearnApp.WebAPI
@@ -30,6 +28,9 @@ namespace LearnApp.WebAPI
 
         public void Configure(IApplicationBuilder app)
         {
+            app.UseDefaultFiles();
+            app.UseStaticFiles();
+
             app.UseRouting();
 
             app.UseEndpoints(endpoints =>
